Data can be pushed to serial port now
Found a bug: an exception occurs when a lot of "data arrived" events are
occured. I think, that reason is in multithreading, because serial port
event listener is running in it's own thread. When a lot of data arrives
from port, the listener begun to spam with "data arrived" events, and
the program crashes.